src/sample_cpp/standalone/requirements.txt
conan==1.58.0
